Refactor `ConstrainedResizable` component
